# DOCS_ARCHIVE/docs/release/universal-review-engine/api/review-engine-openapi.yaml
openapi: 3.1.0
info:
  title: Moshimoshi Review Engine API
  description: |
    Universal Review Engine API for managing spaced repetition learning, content pinning, 
    and review sessions. Supports offline-first architecture with sync capabilities.
  version: 1.0.0
  contact:
    name: Moshimoshi Development Team
    email: api@moshimoshi.app
  license:
    name: Proprietary
servers:
  - url: https://api.moshimoshi.app/api/review/v1
    description: Production server
  - url: https://staging-api.moshimoshi.app/api/review/v1
    description: Staging server
  - url: http://localhost:3000/api/review/v1
    description: Development server
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /sessions/{sessionId}:
    get:
      tags:
        - Session
      summary: Get session status
      description: Retrieve current session state and progress
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
          
    post:
      tags:
        - Session
      summary: Submit answer
      description: Submit an answer for the current review item
      operationId: submitAnswer
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Answer processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      description: Unique session identifier
      schema:
        type: string
        format: uuid
  schemas:
    ReviewableContent:
      type: object
      required:
        - id
        - contentType
        - primaryDisplay
        - primaryAnswer
      properties:
        id:
          type: string
        contentType:
          type: string
          enum: [kana, kanji, vocabulary, sentence, custom]
        primaryDisplay:
          type: string
        secondaryDisplay:
          type: string
        tertiaryDisplay:
          type: string
        primaryAnswer:
          type: string
        alternativeAnswers:
          type: array
          items:
            type: string
        audioUrl:
          type: string
          format: uri
        imageUrl:
          type: string
          format: uri
        difficulty:
          type: number
          minimum: 0
          maximum: 1
        tags:
          type: array
          items:
            type: string
        supportedModes:
          type: array
          items:
            type: string
            enum: [recognition, recall, listening]
        srsData:
          $ref: '#/components/schemas/SRSData'
        pinned:
          type: boolean
    SRSData:
      type: object
      properties:
        interval:
          type: integer
          description: Days until next review
        easeFactor:
          type: number
          minimum: 1.3
          maximum: 2.5
        repetitions:
          type: integer
        lastReviewedAt:
          type: string
          format: date-time
        nextReviewAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [new, learning, review, mastered]
...
    SessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, paused, completed, abandoned]
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewableContent'
        currentIndex:
          type: integer
        totalItems:
          type: integer
        progress:
          type: object
          properties:
            completed:
              type: integer
            correct:
              type: integer
            incorrect:
              type: integer
            accuracy:
              type: number
            streak:
              type: integer
        estimatedTime:
          type: integer
          description: Estimated minutes remaining
        startedAt:
          type: string
          format: date-time
    SubmitAnswerRequest:
      type: object
      required:
        - itemId
        - answer
        - responseTime
      properties:
        itemId:
          type: string
        answer:
          type: string
        responseTime:
          type: integer
          description: Response time in milliseconds
        confidence:
          type: integer
          minimum: 1
          maximum: 5
        hintsUsed:
          type: integer
    AnswerResponse:
      type: object
      properties:
        correct:
          type: boolean
        expectedAnswer:
          type: string
        feedback:
          type: string
        itemUpdate:
          $ref: '#/components/schemas/SRSData'
        sessionProgress:
          type: object
          properties:
            current:
              type: integer
            total:
              type: integer
            accuracy:
              type: number
            streak:
              type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...